in test_wsloop.py, add end-to-end test that mimics what "tsp run" and "tsp plot" do on wsloop. Fix as needed
